Created a separate group for data models
