+++ ImageFeedApp/Models/OAuthTokenResponseBody.swift
@@ -1,0 +1,22 @@
+//
+//  OAuthTokenResponseBody.swift
+//  ImageFeedApp
+//
+//  Created by Vlad Vintenbakh on 3/2/24.
+//
+
+import Foundation
+
+struct OAuthTokenResponseBody: Codable {
+    let accessToken: String
+    let tokenType: String
+    let scope: String
+    let createdAt: Int
+    
+    enum CodingKeys: String, CodingKey {
+        case accessToken = "access_token"
+        case tokenType = "token_type"
+        case scope
+        case createdAt = "created_at"
+    }
+}
