+++ ImageFeedApp/Profile/ProfileService.swift
@@ -47,19 +47,3 @@ final class ProfileService {
         task.resume()
     }
 }
-
-extension ProfileService {
-    private struct ProfileResult: Codable {
-        let username: String
-        let firstName: String
-        let lastName: String?
-        let bio: String?
-        
-        enum CodingKeys: String, CodingKey {
-            case username
-            case firstName = "first_name"
-            case lastName = "last_name"
-            case bio
-        }
-    }
-}
