+++ ImageFeedApp/ServiceAccessLayer/OAuth2/OAuth2Service.swift
@@ -62,18 +62,4 @@ extension OAuth2Service {
         let url = URL(string: "https://unsplash.com")
         return URLRequest.makeHTTPRequest(path: pathString, httpMethod: "POST", baseURL: url!)
     }
-    
-    private struct OAuthTokenResponseBody: Codable {
-        let accessToken: String
-        let tokenType: String
-        let scope: String
-        let createdAt: Int
-        
-        enum CodingKeys: String, CodingKey {
-            case accessToken = "access_token"
-            case tokenType = "token_type"
-            case scope
-            case createdAt = "created_at"
-        }
-    }
 }
